refactor: move types.ts to utils folder
