+++ together-wherever/src/app/discover/DiscoverPageContent.tsx
@@ -4,7 +4,7 @@ import {useState, useEffect} from "react";
 import {useParams} from "next/navigation";
 import PlaceSearchBox from "@/app/components/PlaceSearchBox";
 import MapView from "@/app/components/Map";
-import {PlaceDetails} from "@/app/types";
+import {PlaceDetails} from "@/app/utils/types";
 import axios from "axios";
 import {MapPinIcon, PhoneIcon, StarIcon, CheckIcon} from "@heroicons/react/24/solid";
 import {NearbyCard} from "@/app/components/cards/NearbyCard";
