+++ together-wherever/src/app/discover/page.tsx
@@ -3,7 +3,7 @@
 import {useState} from "react";
 import PlaceSearchBox from "@/app/components/PlaceSearchBox";
 import MapView from "@/app/components/Map";
-import {PlaceDetails} from "@/app/types";
+import {PlaceDetails} from "@/app/utils/types";
 import axios from "axios";
 import {MapPinIcon, PhoneIcon, StarIcon, CheckIcon} from "@heroicons/react/24/solid";
 import {NearbyCard} from "@/app/components/cards/NearbyCard";
